move tabs into the header
